First Fit function is up and running, plus other minor changes
AddrSpace.cc, Addrspace.h:

I made the FirstFit, WorstFit, and BestFit functions public members of
the AddrSpace class in order to access numPages, which holds the amount
of pages the program will need. NumPages is a private member of the
AddrSpace class.

I had to change the memory allocation function to a public function of
the AddrSpace class in order to call FirstFit, BestFit, and WorstFit

main.cc:

I commented out the line in main.cc that made MemoryAlocation() a global
function.
The global function could not call FirstFit,WorstFit, and BestFit once
they were
members of the AddrSpace class.Now it will not be able to be called in
main.cc as a debugger tool. I commented out the debug lines in main.cc
that were calling MemoryAllocation, since their references are now
undefined